# Various performance improvements and minor tweaks