--- a/src/Colourful/Conversion/ColourfulConverter.ToLab.cs
+++ b/src/Colourful/Conversion/ColourfulConverter.ToLab.cs
@@ -65,7 +65,7 @@ namespace Colourful.Conversion
         {
             if (color == null) throw new ArgumentNullException(nameof(color));
 
-            // conversion (perserving white point)
+            // conversion (preserving white point)
             var converter = new LChabToLabConverter();
             var unadapted = converter.Convert(color);