--- a/src/Colourful/Difference/CMCColorDifference.cs
+++ b/src/Colourful/Difference/CMCColorDifference.cs
@@ -117,6 +117,9 @@ namespace Colourful.Difference
         }
     }
 
+    /// <summary>
+    /// Weighting parameters for CMC l:c
+    /// </summary>
     public enum CMCColorDifferenceThreshold
     {
         /// <summary>