--- a/src/Colourful/Implementation/Conversion/RGB/LinearRGBToXYZConverter.cs
+++ b/src/Colourful/Implementation/Conversion/RGB/LinearRGBToXYZConverter.cs
@@ -52,10 +52,7 @@ namespace Colourful.Implementation.Conversion
 
             var xyz = _conversionMatrix.MultiplyBy(input.Vector);
 
-            double x, y, z;
-            xyz.AssignVariables(out x, out y, out z);
-
-            var converted = new XYZColor(x, y, z);
+            var converted = new XYZColor(xyz);
             return converted;
         }