--- a/src/Colourful/Other/CCTConverter.cs
+++ b/src/Colourful/Other/CCTConverter.cs
@@ -14,6 +14,7 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Text;
+using Colourful.Implementation;
 
 namespace Colourful
 {
@@ -34,21 +35,21 @@ namespace Colourful
             double x_c;
 
             if (temperature <= 4000) // correctly 1667 <= T <= 4000
-                x_c = -0.2661239*(Math.Pow(10, 9)/Math.Pow(temperature, 3)) - 0.2343580*(Math.Pow(10, 6)/Math.Pow(temperature, 2)) + 0.8776956*(Math.Pow(10, 3)/temperature) + 0.179910;
+                x_c = -0.2661239*(1000000000/MathUtils.Pow3(temperature)) - 0.2343580*(1000000/MathUtils.Pow2(temperature)) + 0.8776956*(1000/temperature) + 0.179910;
 
             else // correctly 4000 <= T <= 25000
-                x_c = -3.0258469*(Math.Pow(10, 9)/Math.Pow(temperature, 3)) + 2.1070379*(Math.Pow(10, 6)/Math.Pow(temperature, 2)) + 0.2226347*(Math.Pow(10, 3)/temperature) + 0.240390;
+                x_c = -3.0258469*(1000000000/MathUtils.Pow3(temperature)) + 2.1070379*(1000000/MathUtils.Pow2(temperature)) + 0.2226347*(1000/temperature) + 0.240390;
 
             double y_c;
 
             if (temperature <= 2222) // correctly 1667 <= T <= 2222
-                y_c = -1.1063814*Math.Pow(x_c, 3) - 1.34811020*Math.Pow(x_c, 2) + 2.18555832*x_c - 0.20219683;
+                y_c = -1.1063814*MathUtils.Pow3(x_c) - 1.34811020*MathUtils.Pow2(x_c) + 2.18555832*x_c - 0.20219683;
 
             else if (temperature <= 4000) // correctly 2222 <= T <= 4000
-                y_c = -0.9549476*Math.Pow(x_c, 3) - 1.37418593*Math.Pow(x_c, 2) + 2.09137015*x_c - 0.16748867;
+                y_c = -0.9549476*MathUtils.Pow3(x_c) - 1.37418593*MathUtils.Pow2(x_c) + 2.09137015*x_c - 0.16748867;
 
             else // correctly 4000 <= T <= 25000
-                y_c = +3.0817580*Math.Pow(x_c, 3) - 5.87338670*Math.Pow(x_c, 2) + 3.75112997*x_c - 0.37001483;
+                y_c = +3.0817580*MathUtils.Pow3(x_c) - 5.87338670*MathUtils.Pow2(x_c) + 3.75112997*x_c - 0.37001483;
 
             return new xyChromaticityCoordinates(x_c, y_c);
         }