… src/Colourful/Implementation/Utils/Extensions.cs
… src/Colourful/Implementation/Utils/MathUtils.cs