This pull request addresses several performance bottlenecks (#18) and brings some other minor tweaks like spelling fixes and refactoring.

Biggest gains come from rewriting matrix-vector multiplication and replacing calls of Math.Pow where the exponent is an integer. In the following some micro-benchmarking results for selected conversion and difference methods:

|                   Method |     Elapsed |  Allocated |    Elapsed (new) | Allocated (new) |
|------------------------- |------------:|-----------:|------------:|-----------:|
|                 RGBToLab | 13,266.9 ns |     5528 B | 2,633.49 ns |     1344 B |
|                 LabToRGB | 12,810.7 ns |     5664 B | 2,979.04 ns |     1568 B |
|                 XYZToLab |  9,846.5 ns |     4352 B | 2,149.70 ns |      992 B |
|                 LabToXYZ |  9,429.8 ns |     4344 B | 1,749.33 ns |      984 B |
|                 RGBToXYZ |  2,835.2 ns |     1176 B |   700.28 ns |      352 B |
|                 XYZToRGB |  3,300.2 ns |     1320 B | 1,251.77 ns |      584 B |
|     CIE76ColorDifference |    247.3 ns |      296 B |    11.25 ns |        0 B |
|     CIE94ColorDifference |    260.0 ns |        0 B |    42.65 ns |        0 B |
| CIEDE2000ColorDifference |  1,235.5 ns |      280 B |   611.14 ns |        0 B |
|       CMCColorDifference |    181.9 ns |        0 B |   198.41 ns |        0 B |