@tompazourek 2017-11-04
This looks great, thanks for the PR. I'm away now, but I'll check this out in the next few days. I'm looking forward to learn about those optimizations :)